move norm test to own file
